# swagger-app/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'API for Mentoring'
  description: |
    Homework 8 (deadline 12.12.17)
    [Repository on GitHub](https://github.com/Zagula/node-js-mentoring)
  contact:
    email: Andrei_Tsiarnouski@epam.com
tags:
- name: products
  description: Manipulating with products
- name: users
  description: Manipulating with users
- name: cities
  description: Manipulating with cities
paths:
  /products:
    get:
      tags:
      - products
      responses:
        200:
          description: Return ALL products
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
    post:
      tags:
      - products
      parameters:
      - in: query
        name: name
        type: string
        description: The name of new product
        required: true
      - in: query
        name: brand
        type: string
        description: The name of new product's brand
      - in: query
        name: price
        type: number
        description: The price of new product
      - in: query
        name: color
        type: string
        description: The color of new product
      - in: query
        name: size
        type: string
        description: The size of new product
      responses:
        200:
          description: Return created product
          schema:
            $ref: '#/definitions/Product'
  /products/{productId}:
    get:
      tags:
      - products
      parameters:
      - name: productId
        in: path
        type: string
        required: true
        description: Product ID for details
      responses:
        200:
          description: Return product with corresponding Id
          schema:
            $ref: '#/definitions/Product'
    delete:
      tags:
      - products
      parameters:
      - name: productId
        type: string
        in: path
        required: true
      responses:
        200:
          description: Product deleted successfully
        400:
          description: Wrong ID
        404:
          description: Product was not found
  /products/{productId}/reviews:
    get:
      tags:
      - products
      parameters:
      - name: productId
        type: string
        in: path
        required: true
      responses:
        200:
          description: Return ALL product's reviews
          schema:
            type: array
            items:
              $ref: '#/definitions/Review'
  /users:
    get:
      tags:
      - users
      responses:
        200:
          description: Return ALL users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
  /users/{userId}:
    delete:
      tags:
      - users
      parameters:
      - name: userId
        type: string
        in: path
        required: true
      responses:
        200:
          description: User deleted successfully
        400:
          description: Wrong ID
        404:
          description: User was not found
  /cities:
    get:
      tags:
      - cities
      responses:
        200:
          description: Return all cities
          schema:
            type: array
            items:  
              $ref: '#/definitions/City'
    post:
      tags:
      - cities
      parameters:
      - name: data
        in: body
        schema:
          $ref: '#/definitions/City'
      responses:
        200:
          description: Return created city
          schema:
            $ref: '#/definitions/City'
  /cities/{cityId}:
    put:
      tags:
      - cities
      parameters:
      - name: cityId
        in: path
        type: string
        required: true
      - name: data
        in: body
        schema:
          $ref: '#/definitions/City'
      responses:
        200:
          description: Return updated (or created) city
          schema:
            $ref: '#/definitions/City'
    delete:
      tags:
      - cities
      parameters:
      - name: cityId
        type: string
        in: path
        required: true
      responses:
        200:
          description: City deleted successfully
        400:
          description: Wrong ID
        404:
          description: City was not found
definitions:
  ProductOptions:
    type: object
    properties:
      color:
        type: string
        example: gray
      size:
        type: string
        example: XL
  Product:
    type: object
    required:
    - name
    properties:
      name:
        type: string
        example: Commom Jeans
      brand:
        type: string
        example: China Mall
      price:
        type: integer
        example: 27.99
      options:
        $ref: '#/definitions/ProductOptions'
  Location:
    type: object
    properties:
      lat:
        type: integer
        example: 30.5678
      long:
        type: integer
        example: 50.1234
  City:
    type: object
    properties:
      name:
        type: string
        example: Gomel
      capital:
        type: boolean
        example: false
      country:
        type: string
        example: Belarus
      location:
        $ref: '#/definitions/Location'
  Review:
    type: object
    properties:
      productId:
        type: number
        example: 4
      userId:
        type: number
        example: 2
      review:
        type: string
        example: The best purchase ever !!!!
  User:
    type: object
    properties:
      name:
        type: string
        example: John
      age:
        type: number
        example: 45
# Added by API Auto Mocking Plugin
host: 127.0.0.1:8000
basePath: /api
schemes:
 - http